Fix fftpack vs fftw tests.
Add another real to real normalization test.

Skip some tests if scipy is not installed.

Most of the binding tests require scipy.

Undo the last commit.

SciPy only supports three.

Overhaul the scipy tests.

np -> numpy

Match the convention elsewhere in the test suite.

Clean up the real to real tests.

They now conform, more or less, to the rest of the tests in terms of
their usage of infrastructure.

Get rid of print statements.

Convert some unused verification code into a test.

This test verifies that all of the lookup tables (which are used through
out the real to real tests) to the right thing.