Provide a Dask interface to pyFFTW's 1-D FFTs
Wrap the existing NumPy interface so that it will integrate nicely with
Dask. As Dask currently only supports 1-D FFT functions, simply wrap and
provide those for now. In the future Dask will support 2-D and N-D FFT
function. This will be updated at that time to provide those too. Also
add borrow tests from the NumPy interface to use with the Dask
interface. Update it to use the Dask interface, work with Dask Arrays,
and drop any tests that deal with mutation or 2-D/N-D FFT functions.
Further autdoc all functions in the interface so as to include them in
the documentation.