ENH: add next_fast_len utility function
This function recommends the next largest efficient FFT size.

It is based on scipy's routine of the same name (credit to @endolith and the
scipy devs).

FFTW allows additional prime factors over scipy's fftpack.
Specifically, the FFTW3 documentation recommend sizes that are:

2**a * 3**b * 5**c * 7**d * 11**e * 13**f where e+f is either 0 or 1

As this routine has utility outside of the scipy interfaces it is defined
in pyfftw/utils.pxi and imported from there into the scipy interfaces.

renamed hams variable to lpre as these precomputed lengths are no longer
all Hamming numbers.