Use `pyfftw.version` to get version information
Relying on the `setup.py` file to get the version information doesn't
seem like a good idea. It also requires `pyfftw` to be built in source
to build the docs, which really shouldn't be a requirement. Given
`pyfftw.version` will have this information and should be created in
source or out-of-source builds, use that instead to extract the version
information.