Move docs out of the source code
Previously the docs were embedded with the source code. This is not a
particularly common practice when working with Sphinx. Instead most
projects prefer to have the docs in a separate directory. We restructure
the docs in this commit to also move them out of the source code and
into a separate directory.