Make the home url the GitHub repo
Instead of pointing to GitHub pages in the package metadata, simply
point to the GitHub repo. Currently the GitHub pages are out-of-date and
better documentation can be found on Read the Docs. Given this can
easily be found from the Readme in the repo. Simply direct people to the
repo instead.