Fix Dask interface test name
We are testing the Dask interface, not the NumPy one. So update the
interface argument name to match.

Also as the Dask interface just takes `None` for all of its arguments,
it's better to inspect the arguments of the NumPy interface that it
wraps, which actually take non-trivial arguments. Hence the tests are
updated to use this strategy.