Validate the Dask FFT interface exists
In some cases it appears Dask is installed, it is new enough, and has
support for the FFT wrapper. However there is no guarantee that a new
version of Dask will work on older Pythons. For instance, Python 3.3, is
no longer tested upstream and has not been for a while. Also Python 3.4
only sees some testing on `master` not PRs. Not to mention the
dependencies pulled in may or may not work with Dask on these older
Python versions. As a result it appears the Dask FFT interface may be
missing. So instead of testing proxies to see if Dask is installed and
functional, simply verify the Dask FFT interface for pyFFTW exists. If
not, skip all related tests.