Simply check `fft_wrap` for Dask interface
The only important constraint for determine if the Dask interface should
be loaded is to check whether `fft_wrap` exists. If it does, we can wrap
up the pyFFTW functions without issues. However if `fft_wrap` does not
exist, we should not load the Dask interface as it will not be possible
to wrap our FFT functions for Dask to use.