Assign `dask_fft` a dummy value if it's missing
Some of the code in the `InterfacesDaskFFTTestFFT` class that is defined
as part of the class definition (as opposed to a method) is run whether
or not the class is skipped. As such, we assign a dummy value to
`interfaces.dask_fft` so this code can run through untouched. However as
the tests are still skipped in this case, the assignment itself has no
other effect.