MNT: Bump Dask minimum to 0.15.0
Support other miscellaneous FFT functions added to Dask. Namely FFT
shift and frequency functions. These are mainly imported as a
convenience to the user and are not strictly necessary. Since these were
added in Dask 0.15.0 and were not available previously, bump the Dask
version so we know these functions are included.