MNT: Include Dask interface only for NumPy 1.10+
As the Dask versions we are interested in are not really compatible with
any NumPy before 1.10+ and can sometimes fail (even on import) in
bizarre ways, simply add a version check to ensure that NumPy 1.10+ is
installed before even trying to import Dask. Make sure to clean up
everything imported whether or not the Dask interface can be added
independently of this NumPy version check.