MNT: Import Dask in tests if interface exists
Since we already check for the Dask interface and can see if it is not
present, simply use that as a guide to whether Dask can be imported in
the test suite. Skip the imports if the Dask interface does not exist.
Store dummy values for all skipped imports as some of them are used in
locations that are difficult to skip with the test suite (e.g. class
definitions). This allows that code to run even though all associated
tests that would use it are already skipped.