MNT: Add optional SciPy requirement
Includes an optional requirement on SciPy. As our oldest version of
SciPy, 0.12.0, requires NumPy 1.5.0 at a minimum to work and our minimum
supported NumPy version is 1.6.0, our SciPy support should work fine
with our existing hard NumPy requirement. Further it is expected modern
versions of SciPy would use more recent packaging standards to affect
its NumPy version requirement. Thus making it unnecessary for us to
specify it. Not to mention SciPy tries to support much older versions of
NumPy than we or other developers are likely to support. So there
shouldn't be any need for this specification as there was for Dask.