Optionally use Dask if Numpy is 1.10
The previous Dask interface import guard check required NumPy to be
greater than 1.10.0 to work. However it is also ok for NumPy to be equal
to 1.10.0. This relaxes the check a bit to allow NumPy 1.10.0 for
instance. However filters out dev and rc releases. The latter should
have no practical importance given how old NumPy 1.10 is. Nevertheless
it is good to be strict about having a formal release to avoid potential
issues.