release notes for 0.11 (#188)
* DOC: add release notes for 0.11.0

* Update release notes to reflect recently merged PRs

* Update general notes about Dask interface

Point out the Dask FFT interface is a drop-in replacement for the
equivalent module in Dask.

* Updates to the Dask interface release note section

Notes that the Dask interface is a drop-in replacement for the Dask
Array FFT module for all dimensionalities. Mentions other functions are
imported as a convenience to the user.

* Fix missing tick in issue list

* Add some recent issues and PRs to the list

Various bug fixes mostly targeted towards building and installing. Some
other bug fixes as well.

* Add Dask issues/PRs

Mainly focused on N-D support with some test fixes and better handling
of it being an optional dependency.

* Other misc PRs and issues

Mainly related to CIs.

* update release notes to reflect recent changes