Update Dask FFT listing in FFT interface overview  [ci skip] (#254)
On the interface overview page, FFT interfaces for NumPy, SciPy, and
Dask are listed. The Dask listing on this page was missing the 2-D and
N-D functions that have been added since. This updates the listing on
the overview page to include these 2-D and N-D Dask FFT interface
functions.