Remove conditional compilation surrounding function definitions.
As a replacement the functions are always defined, but we
conditionally add macros which hide the underlying FFTW calls in cases
where these are not available (ensuring that the missing symbols
aren't used). The empty Cython wrapper functions will not actually be
called in normal usage since the same preprocessor definitions are
used both for defining the replacement macros and for selecting
whether the functions are used at runtime.